CyberPunkCalc
Cyberpunkcity is a calculator written in Python, with the addition of the Kivi library, which allows you to further convert the application into an APK file and use it on any Android device.

Be careful with the sound, a soundtrack has been added to the calculator so that it is not boring to count.

Implemented functions: division, multiplication, addition and subtraction. Also, the ".\0." function has been added, which allows you not to press "Zero" and "Dot" separately. if you are writing a new number that is decimal, just press the " .\0." key. Count and save time!